Review fixes.

+++ search/common.hpp
@@ -13,7 +13,8 @@ namespace search
 //              the prefix and non-prefix tokens.
 using QueryTokens = buffer_vector<strings::UniString, 32>;
 
-using Locales = base::SafeSmallSet<CategoriesHolder::kMaxSupportedLocaleIndex + 1>;
+using Locales =
+    base::SafeSmallSet<static_cast<uint64_t>(CategoriesHolder::kMaxSupportedLocaleIndex) + 1>;
 
 /// Upper bound for max count of tokens for indexing and scoring.
 int constexpr MAX_TOKENS = 32;
